# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SOAPP Documentación",
        "description": "Salud Ocupacional APP",
        "contact": {
            "email": "daruiza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8080/api",
            "description": "SOAPP API Server"
        }
    ],
    "paths": {
...
        "/report/index": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Get All Reports",
                "description": "Return Reports",
                "operationId": "getAllReports",
                "parameters": [
                    {
                        "name": "commerce_id",
                        "in": "query",
                        "description": "Report commerce_id",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Report year",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "month",
                        "in": "query",
                        "description": "Report month",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/report/store": {
            "post": {
                "tags": [
                    "Report"
                ],
                "summary": "Store Report",
                "description": "Store Report",
                "operationId": "storeReport",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Report"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/report/showbyreportid/{id}": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Get One Report By one Id",
                "description": "Return One Report",
                "operationId": "getReportById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/report/update/{id}": {
            "put": {
                "tags": [
                    "Report"
                ],
                "summary": "Update One Report By one Id",
                "description": "Update One Report",
                "operationId": "getUpdateReportById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Report"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/report/destroy/{id}": {
            "delete": {
                "tags": [
                    "Report"
                ],
                "summary": "Delete One Report By one Id",
                "description": "Delete One Report",
                "operationId": "getDestroyReportById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Report Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Report": {
                "title": "Report",
                "description": "Report body data",
                "properties": {
                    "project": {
                        "title": "project",
                        "description": "project of the Report",
                        "type": "string",
                        "example": "Caida"
                    },
                    "responsible": {
                        "title": "responsible",
                        "description": "Responsible of the Report",
                        "type": "string",
                        "example": "CAIDAD"
                    },
                    "email_responsible": {
                        "title": "email_responsible",
                        "description": "email_responsible of the Report",
                        "type": "string",
                        "example": "david@gmail.com"
                    },
                    "phone_responsible": {
                        "title": "phone_responsible",
                        "description": "phone_responsible of the Report",
                        "type": "string",
                        "example": "123456789"
                    },
                    "date": {
                        "title": "date",
                        "description": "date of the Report",
                        "type": "string",
                        "format": "date",
                        "example": "2020-01-01"
                    },
                    "commerce_id": {
                        "title": "commerce_id",
                        "description": "commerce_id of the Report",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}